Port block_scaling_offset

+++ blocklib/digitizers/median_and_average/median_and_average_cpu.cc
@@ -56,9 +56,9 @@ work_return_t median_and_average_cpu<T>::work(work_io &wio) {
     const auto in      = wio.inputs()[0].items<T>();
     auto       out     = wio.outputs()[0].items<T>();
 
-    const auto med_len = pmtf::get_as<std::size_t>(*this->param_n_med);
-    const auto avg_len = pmtf::get_as<std::size_t>(*this->param_n_lp);
-    const auto vec_len = pmtf::get_as<std::size_t>(*this->param_vec_len);
+    const auto med_len = static_cast<int>(pmtf::get_as<std::size_t>(*this->param_n_med));
+    const auto avg_len = static_cast<int>(pmtf::get_as<std::size_t>(*this->param_n_lp));
+    const auto vec_len = static_cast<int>(pmtf::get_as<std::size_t>(*this->param_vec_len));
 
     float      buffer[2 * med_len + 1];
     float      temp_buffer[vec_len];
